+++ dds_cli/auth.py
@@ -34,6 +34,7 @@ class Auth(base.DDSBaseClass):
         force_renew_token: bool = True,  # Only used if authenticate is True
         token_path: str = None,
         totp: str = None,
+        allow_group: bool = False,
     ):
         """Handle actions regarding session management in DDS."""
         # Initiate DDSBaseClass to authenticate user
@@ -43,6 +44,7 @@ class Auth(base.DDSBaseClass):
             force_renew_token=force_renew_token,
             token_path=token_path,
             totp=totp,
+            allow_group=allow_group,
         )
 
     def check(self):
@@ -85,7 +87,6 @@ class Auth(base.DDSBaseClass):
 
             response_json, _ = dds_cli.utils.perform_request(
                 endpoint=dds_cli.DDSEndpoint.USER_ACTIVATE_HOTP,
-                headers=None,
                 method="post",
                 auth=(username, password),
             )